Fix reference unittest for timm
